requirements.txt
django
djangorestframework
psycopg2-binary
django-crispy-forms
requests
celery[redis]
django-celery-beat
django-celery-results